refactor(storage): update isBrowser check and move to env module
- Update `isBrowser` check to use a function instead of a constant.
- Move `isBrowser` from `utils` to `env` module.
- Update import paths and usage in various files.
- Rename `utils.test.ts` to `env.test.ts` and update tests accordingly.

This change ensures that the `isBrowser` check is more flexible and can be used in different environments. It also organizes the code by moving environment-related checks to a dedicated `env` module.
