+++ packages/cosec/test/inMemoryStorage.test.ts
@@ -12,8 +12,7 @@
  */
 
 import { describe, expect, it, vi } from 'vitest';
-import { getStorage, InMemoryStorage } from '../src/inMemoryStorage';
-import * as utils from '../src/utils';
+import { getStorage, InMemoryStorage } from '../src';
 
 describe('inMemoryStorage.ts', () => {
   describe('InMemoryStorage', () => {
@@ -139,10 +138,8 @@ describe('inMemoryStorage.ts', () => {
         key: vi.fn(),
         length: 0,
       };
-      
-      // Spy on isBrowser and make it return true
-      const isBrowserSpy = vi.spyOn(utils, 'isBrowser', 'get').mockReturnValue(true);
-      
+
+
       // Temporarily set window.localStorage
       const originalWindow = (globalThis as any).window;
       (globalThis as any).window = {
@@ -155,8 +152,6 @@ describe('inMemoryStorage.ts', () => {
       } finally {
         // Restore original window
         (globalThis as any).window = originalWindow;
-        // Restore spy
-        isBrowserSpy.mockRestore();
       }
     });
   });
